+++ frontend/src/Configuration/Configuration.tsx
@@ -15,16 +15,7 @@ import {
   Theme,
   Typography,
 } from "@material-ui/core";
-import {
-  mdiAccessPoint,
-  mdiAccount,
-  mdiKey,
-  mdiLock,
-  mdiProtocol,
-  mdiRestart,
-  mdiRocketLaunch,
-  mdiTimerOutline,
-} from "@mdi/js";
+import { mdiRestart } from "@mdi/js";
 import Icon from "@mdi/react";
 
 import { parsedQuery } from "../Utils";
@@ -32,122 +23,7 @@ import { useSettings } from "../Utils";
 import { WebSocketConnection } from "../Common/WebSocket";
 import Footer from "../Common/Footer";
 import Header from "../Common/Header";
-import Section, { ConfigurationSection } from "./Section";
-
-export interface Configuration {
-  [section: string]: ConfigurationSection;
-}
-
-export const defaultConfiguration: Configuration = {
-  general: {
-    name: "General",
-    items: {
-      launchOnStartup: {
-        name: "Launch on Startup",
-        description: "Start the application on startup.",
-        defaultValue: false,
-        value: null,
-        icon: mdiRocketLaunch,
-      },
-    },
-  },
-  network: {
-    name: "Network",
-    description: "API and WebSocket specific settings.",
-    items: {
-      apiPort: {
-        name: "API Port",
-        description: "The port the application's API runs on.",
-        defaultValue: 9170,
-        value: null,
-        icon: mdiProtocol,
-        minimum: 1,
-        requiresServerRestart: true,
-      },
-      wsPort: {
-        name: "WebSocket Port",
-        description: "The port the WebSocket server runs on.",
-        defaultValue: 9172,
-        value: null,
-        icon: mdiProtocol,
-        minimum: 1,
-        requiresServerRestart: true,
-      },
-      apiKey: {
-        name: "API Key",
-        description: "The API key to authenticate with the API.",
-        defaultValue: "",
-        value: null,
-        icon: mdiKey,
-        requiresServerRestart: true,
-      },
-    },
-  },
-  mqtt: {
-    name: "MQTT",
-    description: "MQTT specific settings.",
-    items: {
-      enabled: {
-        name: "MQTT Enabled",
-        description: "Should MQTT be enabled?",
-        defaultValue: false,
-        value: null,
-        icon: mdiAccessPoint,
-        requiresServerRestart: true,
-      },
-      host: {
-        name: "Broker Host",
-        description: "The host of your MQTT broker.",
-        defaultValue: "localhost",
-        value: null,
-        icon: mdiAccessPoint,
-        requiresServerRestart: true,
-      },
-      port: {
-        name: "Broker Port",
-        description: "The port of your MQTT broker.",
-        defaultValue: 1883,
-        value: null,
-        icon: mdiProtocol,
-        minimum: 1,
-        requiresServerRestart: true,
-      },
-      username: {
-        name: "Broker Username",
-        description: "The username of your MQTT broker.",
-        defaultValue: "",
-        value: null,
-        icon: mdiAccount,
-        requiresServerRestart: true,
-      },
-      password: {
-        name: "Broker Password",
-        description: "The password of your MQTT broker.",
-        defaultValue: "",
-        value: null,
-        icon: mdiLock,
-        isPassword: true,
-        requiresServerRestart: true,
-      },
-    },
-  },
-  observer: {
-    name: "Observer",
-    description: "Observer specific settings.",
-    items: {
-      interval: {
-        name: "Observer Interval (ms)",
-        description:
-          "The amount of time in milliseconds the observer will wait before checking for new data. The faster, the more updates, but also the more system utilization.",
-        defaultValue: 60000,
-        value: null,
-        minimum: 20000,
-        icon: mdiTimerOutline,
-        requiresServerRestart: true,
-      },
-    },
-  },
-};
+import Section from "./Section";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
