+++ frontend/src/entities/application.entity.ts
@@ -1,0 +1,19 @@
+export interface ApplicationInfo {
+  address: string;
+  fqdn: string;
+  host: string;
+  ip: string;
+  mac: string;
+  port: number;
+  updates?: ApplicationUpdate;
+  uuid: string;
+  version: string;
+  websocketAddress: string;
+  websocketPort: number;
+}
+
+export interface ApplicationUpdate {
+  available: boolean;
+  url: string;
+  version: { current: string; new: string };
+}
