+++ frontend/src/entities/event.entity.ts
