+++ frontend/src/entities/settings.entity.ts
@@ -11,3 +11,15 @@ export interface Setting {
   key: string;
   value: string;
 }
+
+export type SettingsValue =
+  | null
+  | boolean
+  | string
+  | number
+  | SettingsObject
+  | SettingsValue[];
+
+export type SettingsObject = {
+  [key: string]: SettingsValue;
+};
